Convert from 'int' to 'long'
When testing
NSCALES=4
NMOMENTS=3
NX=4096
NY=4096
Run test_cleaners.sh, will occur error：
terminate called after throwing an instance of 'std::bad_array_new_length',
The reason is that we need to calculate the value of
nscales * nscales * nmoments * nmoments * ny * nx, which is exceed the
limit of an 'int' type in C++. So I changed the type from 'int' to 'long'

+++ c-kernel/src/image/cleaners.h
@@ -15,10 +15,10 @@ void msmfsclean_kernel(double *m_model, double *residual,
 					   double *scalestack, double *smresidual,
 					   double *ssmmpsf, double *hsmmpsf, double *ihsmmpsf,
 					   double *ldirty, double *psf,
-					   const int nscales, const int nmoments,
-					   const int ny, const int nx,
+					   const long nscales, const long nmoments,
+					   const long ny, const long nx,
 					   double *windowstack,
-					   double gain, double absolutethresh, int niter,
+					   double gain, double absolutethresh, long niter,
 					   const char* findpeak = "CASA");
 
 int msmfsclean_kernel_c(PyArrayObject *&m_model_obj,
@@ -33,7 +33,7 @@ int msmfsclean_kernel_c(PyArrayObject *&m_model_obj,
                  PyArrayObject *&windowstack_obj,
                  double gain,
                  double absolutethresh,
-                 int niter,
+                 long niter,
                  const char* findpeak);
 
 #endif
